\name{baselineCorr}
\alias{baselineCorr}
\title{
Baseline Correction
}
\description{
Offers three different methods for baseline correction of raw spectral data. Methods include monotone minimum, linear interpolation, and loess (1).
}
\usage{
baselineCorr(dat, mass_dat, intensity_dat, method = NULL, n = NULL)
}
\arguments{
  \item{dat}{ The name of the spectral data frame, containing \code{m/z} data in the first column and spectral intensity data in subsequent columns. }
  \item{mass_dat}{ Character string. The name of the column in \code{dat} containing the \emph{m/z} data for the spectrum that will be corrected.  }
  \item{intensity_dat}{ Character string. The name of the column in \code{dat} containing the intensity data for the spectrum to be baseline corrected.  }
  \item{method}{ Character string. The method that is to be used to perform the baseline correction. Either \code{"monotone"} for monotone minimum correction, or \code{"linear"} for linear interpolation. }
  \item{n}{ Single odd numeric value. If \code{method = "linear"}, the size of the window that should be used for linear interpolation of the spectral baseline. }
}
\value{ Returns a new data frame containing the baseline corrected spectral data.
}
\references{
https://github.com/wesleyburr/subMaldi
(1) Yang, C., He, Z. & Yu, W. Comparison of public peak detection algorithms for MALDI mass spectrometry data analysis. BMC Bioinformatics 10, 4 (2009). https://doi.org/10.1186/1471-2105-10-4
}
\author{
Kristen Yeh <kristenyeh@trentu.ca>
Wesley Burr <wburr@trentu.ca>
}

\section{Methods}{
  \describe{
    \item{monotone_min}{ Identifies valleys in the spectrum by evaluating the slopes between adjacent intensity values. Moving from left to right, if slope of point A is greater than 0, the nearest point B with slope of less than 0 is located. Points in between A and B are recorded as potential peaks. If slope of A < 0, nearest point B with slope > 0 is located. Points between A and B are recorded as valleys. Each intensity value in the spectrum is then subtracted by the intensity of the nearest valley, correcting any irregularities in the baseline of the data. }
    \item{linear}{ Divides a spectrum into small segments and evaluates the mean of points in each segment as a valley (1). A baseline is then generated by linearly interpolating valleys across all small segments. In each segment, the intensity of each peak is subtracted by the intensity of the valley in that segment. }
    \item{loess}{
    Divides a spectrum into small segments and evaluates the quantile in each segment. If any intensities in the segment are less than the quantile, these intensities are recorded as baseline predictors. If no intensities are greater than the quantile in the segment, the quantile is recorded as the baseline predictor. The baseline is then estimated by performing local polynomial regression fitting to the predictor points.
    }
    }
    }

\seealso{    
  \code{\link{smoothSpectrum}}
}

\examples{
## Load sample dataset "bsline"
data("bsline")

## Correct the baseline using the monotone minimum method
bsline_mono <- baselineCorr(bsline, "mass", "raw", 
                      method = "monotone_min")

## Correct the baseline using the linear interpolation method
## Window size of 7
bsline_linear <- baselineCorr(bsline, "mass", "raw", 
                      method = "linear", n = 7)
}

\keyword{ methods }
\keyword{ manip }
