<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normSpectra.R
\name{normSpectra}
\alias{normSpectra}
\title{Normalize Spectral Data}
\usage{
normSpectra(
  dat,
  mass_dat,
  method = NULL,
  norm_mz = NULL,
  upper = NULL,
  lower = NULL,
  spectra_cols = NULL,
  showHI = FALSE
)
}
\arguments{
\item{dat}{The name of the spectral data frame, containing \code{m/z} data
in the first column and spectral intensity data in subsequent columns.}

\item{mass_dat}{A character string; the name of the column in \code{dat}
containing the \emph{m/z} data for the spectrum.}

\item{method}{A character string; the normalization method that should be
used to process the data. See 'Methods' below for list of methods. Default
= NULL.}

\item{norm_mz}{Numeric. If \code{method = "custom"}, the \emph{m/z} peak to
which the spectral intensity should be normalized to. Value should have the
same number of decimal places as the \emph{m/z} data in \code{dat}. If
\code{method = "custom_imprecise", this value must be given as a character
string of numbers.}}

\item{upper}{Numeric. If \code{method = "stdev"}, the upper \emph{m/z}
bound of the noise region of the spectrum.}

\item{lower}{Numeric. If \code{method = "stdev"}, the lower \emph{m/z}
bound of the noise region of the spectrum.}

\item{spectra_cols}{A character vector; the names of the column in
\code{dat} containing the intensity data for the spectra-of-interest.}

\item{showHI}{Logical. To be used with \code{method = "custom"}. If
\code{TRUE}, intensity values greater than \code{norm_mz} will be kept. If
\code{FALSE}, any peaks with intensity greater than \code{norm_mz} will be
truncated to the intensity of \code{norm_mz}. Default = FALSE.}
}
\value{
Returns a new data frame including the original \emph{m/z} data and
normalized intensity data.
}
\description{
Normalize spectral data to a common scale using several different methods.
}
\section{Methods}{
 \describe{ \item{max}{ Normalizes the intensity data of
spectra to a scale of 0,1. } \item{max_set}{ Normalizes the intensity data
of spectra to a scale of 0,1, where 1 is the single most intense peak of
the spectral set.} \item{custom}{ Normalizes the intensity data of each
input spectrum to the intensity of the selected \emph{m/z} peak. }
\item{custom_imprecise}{Normalizes the intensity data of each input
spectrum to the intensity of the selected \emph{m/z} peak. Allows for less
precise normalization; if data contains four decimal places, input
\emph{m/z} values can be input to 2 or 3 decimal places.} \item{TIC}{
Evaluates the sum of all intensities (TIC) of each spectrum in a dataset.
If the TIC of all spectra are not equal, their intensities are multiplied
by a normalization factor. } \item{rel_TIC}{ Evaluates the sum of all
intensities (TIC) of each spectrum in a dataset. If the TIC of all spectra
are not equal, their intensities are multiplied by a normalization factor.
Each peak intensity is then divided by the normalized peak intensity so
that each spectrum in the dataset has a TIC of 1.} \item{RMS}{ Normalizes
each spectrum by dividing each intensity by the spectrum's RMS. }
\item{median}{ Evaluates the median of each spectrum in a dataset after
removing 0 values introduced by mapping. If the medians are not equal
between spectra, a normalization factor is applied to the spectral
intensities until all median intensities in the dataset are equal.}
\item{stdev}{ Evaluates the standard deviation of intensity values within
the same noisy region (a region lacking peaks) of each spectrum. All
intensities in each spectrum are then divided by the spectrum's standard
deviation in the noise region. All output spectra should have a standard
deviation of 1 in the selected region. } \item{quantile}{ Normalizes the
distributions of the values in each spectrum in a set.  Sorts the intensity
data of each spectrum and evaluates the average intensity for each rank.
The intensity values are then replaces with the averaged intensities,
rearranged in their original order. } }
}

\examples{

=======
\name{normSpectra}
\alias{normSpectra}
\title{
Normalize Spectral Data
}
\description{
Normalize spectral data to a common scale using several different methods. Capable of normalizing 1-6 spectra at once.
}
\usage{
normSpectra(dat, mass_dat, method = NULL, norm_mz = NULL, 
            upper = NULL, lower = NULL, spec1, spec2 = NULL, 
            spec3 = NULL, spec4 = NULL, spec5 = NULL, spec6 = NULL,
            showHI = FALSE)
}
\arguments{
  \item{dat}{The name of the spectral data frame, containing \code{m/z} data in the first column and spectral intensity data in subsequent columns.}
  \item{mass_dat}{A character string; the name of the column in \code{dat} containing the \emph{m/z} data for the spectrum.}
  \item{method}{A character string; the normalization method that should be used to process the data. See 'Methods' below for list of methods. Default = NULL.}
  \item{norm_mz}{Numeric. If \code{method = "custom"}, the \emph{m/z} peak to which the spectral intensity should be normalized to. Value should have the same number of decimal places as the \emph{m/z} data in \code{dat}.}
  \item{lower}{Numeric. If \code{method = "stdev"}, the lower \emph{m/z} bound of the noise region of the spectrum.}
  \item{upper}{Numeric. If \code{method = "stdev"}, the upper \emph{m/z} bound of the noise region of the spectrum.}
  \item{spec1}{A character string; the name of the column in \code{dat} containing the intensity data for the spectrum to be analyzed.}
  \item{spec2}{A character string; the name of the column in \code{dat} containing the intensity data for the second spectrum to be normalized. Default = NULL.}
  \item{spec3}{A character string; the name of the column in \code{dat} containing the intensity data for the third spectrum to be normalized. Default = NULL.}
  \item{spec4}{A character string; the name of the column in \code{dat} containing the intensity data for the fourth spectrum to be normalized. Default = NULL.}
  \item{spec5}{A character string; the name of the column in \code{dat} containing the intensity data for the fifth spectrum to be normalized. Default = NULL.}
  \item{spec6}{A character string; the name of the column in \code{dat} containing the intensity data for the sixth spectrum to be normalized. Default = NULL.}
  \item{showHI}{Logical. To be used with \code{method = "custom"}. If \code{TRUE}, intensity values greater than \code{norm_mz} will be kept. If \code{FALSE}, any peaks with intensity greater than \code{norm_mz} will be truncated to the intensity of \code{norm_mz}. Default = FALSE.}
}
\details{
}
\value{Returns a new data frame including the original \emph{m/z} data and normalized intensity data.
}
\references{
https://github.com/wesleyburr/subMaldi
}
\author{
Kristen Yeh <kristenyeh@trentu.ca>
Wesley Burr <wburr@trentu.ca>
}
\note{
%%  ~~further notes~~
}

\section{Methods}{
  \describe{
    \item{max}{ Normalizes the intensity data of spectra to a scale of 0,1. }
    \item{max_set}{ Normalizes the intensity data of spectra to a scale of 0,1, where 1 is the single most intense peak of the spectral set.}
    \item{custom}{ Normalizes the intensity data of each input spectrum to the intensity of the selected \emph{m/z} peak. }
    \item{TIC}{ Evaluates the sum of all intensities (TIC) of each spectrum in a dataset. If the TIC of all spectra are not equal, their intensities are multiplied by a normalization factor. }
    \item{rel_TIC}{ Evaluates the sum of all intensities (TIC) of each spectrum in a dataset. If the TIC of all spectra are not equal, their intensities are multiplied by a normalization factor. Each peak intensity is then divided by the normalized peak intensity so that each spectrum in the dataset has a TIC of 1.}
    \item{median}{ Evaluates the median of each spectrum in a dataset after removing 0 values introduced by mapping. If the medians are not equal between spectra, a normalization factor is applied to the spectral intensities until all median intensities in the dataset are equal.}
    \item{stdev}{ Evaluates the standard deviation of intensity values within the same noisy region (a region lacking peaks) of each spectrum. All intensities in each spectrum are then divided by the spectrum's standard deviation in the noise region. All output spectra should have a standard deviation of 1 in the selected region. }
    \item{quantile}{ Normalizes the distributions of the values in each spectrum in a set.  Sorts the intensity data of each spectrum and evaluates the average intensity for each rank. The intensity values are then replaces with the averaged intensities, rearranged in their original order. }
  }
}
\seealso{
\code{\link{normMethod_custom}}, 
\code{\link{normMethod_max}}, 
\code{\link{normMethod_max_set}}, 
\code{\link{normMethod_TIC}},
\code{\link{normMethod_rel_TIC}},
\code{\link{normMethod_median}},
\code{\link{normMethod_stdev}},
\code{\link{normMethod_quantile}}
}
\examples{
>>>>>>> f70720d... TIC norm added
## Load sample dataset "Master2.rda"
data("Master2")

## Normalize spectrum "Before1" to its maximum intensity
<<<<<<< HEAD
<<<<<<< HEAD
ex <- normSpectra(dat = Master2, mass_dat = "full_mz", 
            method = "max", spectra_cols = "Before1")

## Normalize the spectra "Before1" and "Before2" to the TIC
ex <- normSpectra(dat = Master2, mass_dat = "full_mz",
            method = "TIC", spectra_cols = c("Before1", "Before2")

## Normalize spectrum "After1" to the intensity of the peak at m/z 253.22
ex <- normSpectra(dat = Master2, mass_dat = "full_mz", 
            method = "custom", norm_mz = 253.22, spectra_cols = "After1")


}
\references{
https://github.com/wesleyburr/subMaldi
}
\author{
Kristen Yeh <kristenyeh@trentu.ca> Wesley Burr <wburr@trentu.ca>
Sophie Castel <sophie.castel@ontariotechu.net>
}
\keyword{manip}
\keyword{methods}
=======
normSpectra(dat = Master2, mass_dat = "full_mz", method = "max", spec1 = "Before1")
=======
ex <- normSpectra(dat = Master2, mass_dat = "full_mz", 
            method = "max", spec1 = "Before1")
>>>>>>> 6b09f74... Fixed documentation

## Normalize the spectra "Before1" and "Before2" to the TIC
ex <- normSpectra(dat = Master2, mass_dat = "full_mz",
            method = "TIC", spec1 = "Before1", spec2 = "Before2")

## Normalize spectrum "After1" to the intensity of the peak at m/z 253.22
ex <- normSpectra(dat = Master2, mass_dat = "full_mz", 
            method = "custom", norm_mz = 253.22, spec1 = "After1")

}
<<<<<<< HEAD
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
>>>>>>> f70720d... TIC norm added
=======

\keyword{ methods }
\keyword{ manip }
>>>>>>> c6a3769... Updated documentation
