% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectra.R
\name{plotSpectra}
\alias{plotSpectra}
\title{Plot and Compare Spectra}
\usage{
plotSpectra(
  dat,
  mass_dat,
  spectra_cols,
  palette = NULL,
  colours = colorRampPalette(c("#cc6600", "#33ccff"))(length(spectra_cols)),
  span = 5,
  thresh = 0.1,
  lbls = FALSE,
  lbl.fmt = "\%3.4f",
  min_mz = 0,
  max_mz = max(dat[[mass_dat]]),
  min_I = 0,
  max_I = max(dat[spectra_cols]),
  x_ticks = 100,
  nrows = ceiling(length(spectra_cols)/2),
  intensity_scale = "free_y"
)
}
\arguments{
\item{dat}{The name of the spectral data frame, containing the \emph{m/z}
data in the first column.}

\item{mass_dat}{A character string; the name of the column in \code{dat}
containing the \emph{m/z} data.}

\item{spectra_cols}{A character string; the name(s) of the column(s) in \code{dat}
containing the intensity data for the spectra-of-interest.}

\item{palette}{A character element; the RColorBrewer palette to use. See below for available palettes.}

\item{colours}{A character string indicating the desired colour(s)}

\item{span}{Single numeric value; the span of peak maxima in between each
label. Default = 5 (ignores two peak maxima on either side of each label).}

\item{thresh}{Single numeric value (0-100); the threshold of peak
intensities which should be labeled.}

\item{lbls}{Logical. If \code{lbls = TRUE}, labels indicating the
\emph{m/z} value of each peak maxima within the indicated \code{span} will
be included in the output plot. If \code{lbls = FALSE}, labels are not
shown.}

\item{lbl.fmt}{Character string in the format "\%a.bf", where \emph{a} is
the number of figures displayed to the left of decimal places in labels,
and \emph{b} is the number of figures displayed to the right of decimal
places in labels.  Default = "\%3.4f".}

\item{min_mz}{Single numeric value; minimum \emph{m/z} value of the
observed range.}

\item{max_mz}{Single numeric value; upper end of \emph{m/z} range observed
in spectra.}

\item{min_I}{Single numeric value; minimum intensity value of the observed
range.}

\item{max_I}{Single numeric value; upper end of the intensity range
observed in spectra.}

\item{x_ticks}{Single numeric value; the number of ticks on the x-axis.}

\item{nrows}{Single integer value; the number of rows in plot arrangement.}

\item{intensity_scale}{A character string; the method that should be used
for determining the y-axis scales for each spectrum. If \code{method =
"free_y"}, each spectrum will be plotted with its own intensity scale. If
\code{method = "fixed"}, each spectrum will be plotted with the y-axis of
the most intense spectrum in the set.}
}
\value{
Returns a line plot of the input spectra.
}
\description{
Plot single or multiple spectra. For multiple spectra, a grid layout can be called using 'nrows'.
}
\section{RColorBrewer Palettes}{
 
\describe{\itemize{\item{\code{"Accent"}}
                   \item{\code{"Dark2"}} 
                   \item{\code{"Paired"}}  
                   \item{\code{"Pastel1"}}  
                   \item{\code{"Pastel2"}}  
                   \item{\code{"Set1"}}  
                   \item{\code{"Set2"} (default)} 
                   \item{\code{"Set3"}}
                   } 
          }
}

\examples{

## Plotting using the sample dataset "Master.rda"
data("Master")
plotSpectra(dat = Master, mass_dat = "full_mz",
            spectra_cols = c("Blank1", "Blank2"),
            intensity_scale = "free_y", lbls = TRUE, nrows = 2, x_ticks = 10)
}
\references{
https://github.com/wesleyburr/subMaldi
}
\author{
Kristen Yeh <kristenyeh@trentu.ca> Wesley Burr <wburr@trentu.ca> Sophie Castel <sophie.castel@ontariotechu.net>
}
\keyword{aplot}
\keyword{methods}
