<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineCorr.R
\name{baselineCorr}
\alias{baselineCorr}
\title{Baseline Correction}
=======
=======
>>>>>>> 8cb7c44... Baseline correction added
=======
>>>>>>> 4bbf2ad... Baseline correction added
\name{baselineCorr}
\alias{baselineCorr}
\title{
Baseline Correction
}
\description{
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
Offers three different methods for baseline correction of raw spectral data. Methods include monotone minimum, linear interpolation, and LOESS curve fitting.
}
>>>>>>> 4bbf2ad... Baseline correction added
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineCorr.R
\name{baselineCorr}
\alias{baselineCorr}
\title{Baseline Correction}
>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
=======
Offers two different methods for baseline correction of raw spectral data. Methods include monotone minimum and linear interpolation (1).
=======
Offers three different methods for baseline correction of raw spectral data. Methods include monotone minimum, linear interpolation, and loess (1).
>>>>>>> e3eb9d6... Added loess baseline correction
=======
Offers three different methods for baseline correction of raw spectral data. Methods include monotone minimum, linear interpolation, and LOESS curve fitting.
>>>>>>> 2931ed6... Added pre-processing vignette
}
>>>>>>> 8cb7c44... Baseline correction added
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineCorr.R
\name{baselineCorr}
\alias{baselineCorr}
\title{Baseline Correction}
>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
=======
Offers two different methods for baseline correction of raw spectral data. Methods include monotone minimum and linear interpolation (1).
}
>>>>>>> 4bbf2ad... Baseline correction added
\usage{
baselineCorr(dat, mass_dat, intensity_dat, method = NULL, n = NULL)
}
\arguments{
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
=======
>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
=======
>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
\item{dat}{The name of the spectral data frame, containing \code{m/z} data
in the first column and spectral intensity data in subsequent columns.}

\item{mass_dat}{Character string. The name of the column in \code{dat}
containing the \emph{m/z} data for the spectrum that will be corrected.}

\item{intensity_dat}{Character string. The name of the column in \code{dat}
containing the intensity data for the spectrum to be baseline corrected.}

\item{method}{Character string. The method that is to be used to perform
the baseline correction. Either \code{"monotone"} for monotone minimum
correction, \code{"linear"} for linear interpolation, or \code{"loess"} for
LOESS curve fitting.}

\item{n}{Single odd numeric value. If \code{method = "linear"}, the size of
the window that should be used for linear interpolation of the spectral
baseline.}
<<<<<<< HEAD
<<<<<<< HEAD
}
\value{
Returns a new data frame containing the baseline corrected spectral
data.
}
\description{
Offers three different methods for baseline correction of raw spectral
data. Methods include monotone minimum, linear interpolation, and LOESS
curve fitting.
}
\section{Methods}{
 \describe{ \item{monotone_min}{ Identifies valleys in the
spectrum by evaluating the slopes between adjacent intensity values. Moving
from left to right, if slope of point A is greater than 0, the nearest
point B with slope of less than 0 is located. Points in between A and B are
recorded as potential peaks. If slope of A < 0, nearest point B with slope
> 0 is located. Points between A and B are recorded as valleys. Each
intensity value in the spectrum is then subtracted by the intensity of the
nearest valley, correcting any irregularities in the baseline of the data.
} \item{linear}{ Divides a spectrum into small segments and evaluates the
mean of points in each segment as a valley (1). A baseline is then
generated by linearly interpolating valleys across all small segments. In
each segment, the intensity of each peak is subtracted by the intensity of
the valley in that segment. } \item{loess}{ Divides a spectrum into small
segments and evaluates the quantile in each segment. The baseline is
estimated as follows. If the intensity of point A in a segment is less than
the quantile, the intensity of point A is recorded as a baseline predictor
for that point. If the intensity of point A is greater than or equal to the
quantile in the segment, the quantile for the segment is recorded the
baseline predictor for that point. Local polynomial regression fitting is
applied to the predictor points. The baseline is then subtracted from the
raw spectral signal.  } }
}

\examples{

=======
  \item{dat}{ The name of the spectral data frame, containing \code{m/z} data in the first column and spectral intensity data in subsequent columns. }
  \item{mass_dat}{ Character string. The name of the column in \code{dat} containing the \emph{m/z} data for the spectrum that will be corrected.  }
  \item{intensity_dat}{ Character string. The name of the column in \code{dat} containing the intensity data for the spectrum to be baseline corrected.  }
  \item{method}{ Character string. The method that is to be used to perform the baseline correction. Either \code{"monotone"} for monotone minimum correction, \code{"linear"} for linear interpolation, or \code{"loess"} for LOESS curve fitting. }
  \item{n}{ Single odd numeric value. If \code{method = "linear"}, the size of the window that should be used for linear interpolation of the spectral baseline. }
}
\value{ Returns a new data frame containing the baseline corrected spectral data.
=======
>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
}
\value{
Returns a new data frame containing the baseline corrected spectral
data.
}
\description{
Offers three different methods for baseline correction of raw spectral
data. Methods include monotone minimum, linear interpolation, and LOESS
curve fitting.
}
\section{Methods}{
 \describe{ \item{monotone_min}{ Identifies valleys in the
spectrum by evaluating the slopes between adjacent intensity values. Moving
from left to right, if slope of point A is greater than 0, the nearest
point B with slope of less than 0 is located. Points in between A and B are
recorded as potential peaks. If slope of A < 0, nearest point B with slope
> 0 is located. Points between A and B are recorded as valleys. Each
intensity value in the spectrum is then subtracted by the intensity of the
nearest valley, correcting any irregularities in the baseline of the data.
} \item{linear}{ Divides a spectrum into small segments and evaluates the
mean of points in each segment as a valley (1). A baseline is then
generated by linearly interpolating valleys across all small segments. In
each segment, the intensity of each peak is subtracted by the intensity of
the valley in that segment. } \item{loess}{ Divides a spectrum into small
segments and evaluates the quantile in each segment. The baseline is
estimated as follows. If the intensity of point A in a segment is less than
the quantile, the intensity of point A is recorded as a baseline predictor
for that point. If the intensity of point A is greater than or equal to the
quantile in the segment, the quantile for the segment is recorded the
baseline predictor for that point. Local polynomial regression fitting is
applied to the predictor points. The baseline is then subtracted from the
raw spectral signal.  } }
}

\examples{
<<<<<<< HEAD
>>>>>>> 4bbf2ad... Baseline correction added
=======

>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
=======
  \item{dat}{ The name of the spectral data frame, containing \code{m/z} data in the first column and spectral intensity data in subsequent columns. }
  \item{mass_dat}{ Character string. The name of the column in \code{dat} containing the \emph{m/z} data for the spectrum that will be corrected.  }
  \item{intensity_dat}{ Character string. The name of the column in \code{dat} containing the intensity data for the spectrum to be baseline corrected.  }
  \item{method}{ Character string. The method that is to be used to perform the baseline correction. Either \code{"monotone"} for monotone minimum correction, \code{"linear"} for linear interpolation, or \code{"loess"} for LOESS curve fitting. }
  \item{n}{ Single odd numeric value. If \code{method = "linear"}, the size of the window that should be used for linear interpolation of the spectral baseline. }
}
\value{ Returns a new data frame containing the baseline corrected spectral data.
=======
>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
}
\value{
Returns a new data frame containing the baseline corrected spectral
data.
}
\description{
Offers three different methods for baseline correction of raw spectral
data. Methods include monotone minimum, linear interpolation, and LOESS
curve fitting.
}
\section{Methods}{
 \describe{ \item{monotone_min}{ Identifies valleys in the
spectrum by evaluating the slopes between adjacent intensity values. Moving
from left to right, if slope of point A is greater than 0, the nearest
point B with slope of less than 0 is located. Points in between A and B are
recorded as potential peaks. If slope of A < 0, nearest point B with slope
> 0 is located. Points between A and B are recorded as valleys. Each
intensity value in the spectrum is then subtracted by the intensity of the
nearest valley, correcting any irregularities in the baseline of the data.
} \item{linear}{ Divides a spectrum into small segments and evaluates the
mean of points in each segment as a valley (1). A baseline is then
generated by linearly interpolating valleys across all small segments. In
each segment, the intensity of each peak is subtracted by the intensity of
the valley in that segment. } \item{loess}{ Divides a spectrum into small
segments and evaluates the quantile in each segment. The baseline is
estimated as follows. If the intensity of point A in a segment is less than
the quantile, the intensity of point A is recorded as a baseline predictor
for that point. If the intensity of point A is greater than or equal to the
quantile in the segment, the quantile for the segment is recorded the
baseline predictor for that point. Local polynomial regression fitting is
applied to the predictor points. The baseline is then subtracted from the
raw spectral signal.  } }
}

\examples{
<<<<<<< HEAD
>>>>>>> 8cb7c44... Baseline correction added
=======

>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
=======
  \item{dat}{ The name of the spectral data frame, containing \code{m/z} data in the first column and spectral intensity data in subsequent columns. }
  \item{mass_dat}{ Character string. The name of the column in \code{dat} containing the \emph{m/z} data for the spectrum that will be corrected.  }
  \item{intensity_dat}{ Character string. The name of the column in \code{dat} containing the intensity data for the spectrum to be baseline corrected.  }
  \item{method}{ Character string. The method that is to be used to perform the baseline correction. Either \code{"monotone"} for monotone minimum correction, or \code{"linear"} for linear interpolation. }
  \item{n}{ Single odd numeric value. If \code{method = "linear"}, the size of the window that should be used for linear interpolation of the spectral baseline. }
}
\value{ Returns a new data frame containing the baseline corrected spectral data.
}
\references{
https://github.com/wesleyburr/subMaldi
(1) Yang, C., He, Z. & Yu, W. Comparison of public peak detection algorithms for MALDI mass spectrometry data analysis. BMC Bioinformatics 10, 4 (2009). https://doi.org/10.1186/1471-2105-10-4
}
\author{
Kristen Yeh <kristenyeh@trentu.ca>
Wesley Burr <wburr@trentu.ca>
}

\section{Methods}{
  \describe{
    \item{monotone_min}{ Identifies valleys in the spectrum by evaluating the slopes between adjacent intensity values. Moving from left to right, if slope of point A is greater than 0, the nearest point B with slope of less than 0 is located. Points in between A and B are recorded as potential peaks. If slope of A < 0, nearest point B with slope > 0 is located. Points between A and B are recorded as valleys. Each intensity value in the spectrum is then subtracted by the intensity of the nearest valley, correcting any irregularities in the baseline of the data. }
    \item{linear}{ Divides a spectrum into small segments and evaluates the mean of points in each segment as a valley (1). A baseline is then generated by linearly interpolating valleys across all small segments. In each segment, the intensity of each peak is subtracted by the intensity of the valley in that segment. }
    }
    }

\seealso{    
  \code{\link{smoothSpectrum}}
}

\examples{
>>>>>>> 4bbf2ad... Baseline correction added
## Load sample dataset "bsline"
data("bsline")

## Correct the baseline using the monotone minimum method
bsline_mono <- baselineCorr(bsline, "mass", "raw", 
                      method = "monotone_min")

## Correct the baseline using the linear interpolation method
## Window size of 7
bsline_linear <- baselineCorr(bsline, "mass", "raw", 
                      method = "linear", n = 7)
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD

}
\references{
https://github.com/wesleyburr/subMaldi (1) Yang, C., He, Z. &
Yu, W. Comparison of public peak detection algorithms for MALDI mass
spectrometry data analysis. BMC Bioinformatics 10, 4 (2009).
https://doi.org/10.1186/1471-2105-10-4
}
\seealso{
\code{\link{smoothSpectrum}}, \code{\link{peakDet}}
}
\author{
Kristen Yeh <kristenyeh@trentu.ca> Wesley Burr <wburr@trentu.ca>
}
\keyword{manip}
\keyword{methods}
=======
=======
>>>>>>> 8cb7c44... Baseline correction added
=======
>>>>>>> 4bbf2ad... Baseline correction added
}

\keyword{ methods }
\keyword{ manip }
<<<<<<< HEAD
<<<<<<< HEAD
>>>>>>> 4bbf2ad... Baseline correction added
=======

=======

>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
}
\references{
https://github.com/wesleyburr/subMaldi (1) Yang, C., He, Z. &
Yu, W. Comparison of public peak detection algorithms for MALDI mass
spectrometry data analysis. BMC Bioinformatics 10, 4 (2009).
https://doi.org/10.1186/1471-2105-10-4
}
\seealso{
\code{\link{smoothSpectrum}}, \code{\link{peakDet}}
}
\author{
Kristen Yeh <kristenyeh@trentu.ca> Wesley Burr <wburr@trentu.ca>
}
\keyword{manip}
\keyword{methods}
<<<<<<< HEAD
>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
=======
>>>>>>> 8cb7c44... Baseline correction added
=======
>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
=======
>>>>>>> 4bbf2ad... Baseline correction added
