% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpecDF.R
\name{createSpecDF}
\alias{createSpecDF}
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
\title{Create Empty Spectral Data Frame}
=======
%- Also NEED an '\alias' for EACH other topic documented here.
=======
>>>>>>> c6a3769... Updated documentation
=======
>>>>>>> 14c6370... Updated documentation
\title{
Create Empty Spectral Data Frame
}
\description{
Creates an empty data frame to be used for mapping spectral data to a vector of \emph{m/z} values. The vector of \emph{m/z} values is stored in the first column and can be called by "\code{full_mz}".
}
>>>>>>> 270485a... Updated documentation
=======
\title{Create Empty Spectral Data Frame}
>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
=======
\title{Create Empty Spectral Data Frame}
>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
\usage{
createSpecDF(min_mz = 53.76, max_mz = 1100, res = 1e-04, dig = 4)
}
\arguments{
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
\item{min_mz}{Single numeric value; minimum \emph{m/z} value of the
observed range. Default = 53.76.}

\item{max_mz}{Single numeric value; upper end of \emph{m/z} range observed
in spectra. Default = 1100.}

\item{res}{The resolution of peaks; a single numeric value indicating the
step size from the minimum to maximum \emph{m/z} bin. Also known as bin
width. Default = 0.0001.}
=======
=======
>>>>>>> db6ced3... Updated documentation
=======
>>>>>>> 270485a... Updated documentation
  \item{min_mz}{Single numeric value; minimum \emph{m/z} value of the observed range. Default = 53.76.}
  \item{max_mz}{Single numeric value; upper end of \emph{m/z} range observed in spectra. Default = 1100.}
  \item{res}{The resolution of peaks; a single numeric value indicating the step size from the minimum to maximum \emph{m/z} bin. Also known as bin width. Default = 0.0001.}
  \item{dig}{Number of decimal places to round the \emph{m/z} vector to; required for \emph{m/z} values to match full \emph{m/z} vector. Default = 4.}
}
>>>>>>> 270485a... Updated documentation

<<<<<<< HEAD
<<<<<<< HEAD
\item{dig}{Number of decimal places to round the \emph{m/z} vector to;
required for \emph{m/z} values to match full \emph{m/z} vector. Default =
4.}
}
=======
>>>>>>> 5831098... Added pre-processing vignette
=======
>>>>>>> 2931ed6... Added pre-processing vignette
\value{
<<<<<<< HEAD
Returns a data frame which can be used to map irregularly spaced
spectral data to a set range of \emph{m/z} values, contained in the first
column of the frame.
}
<<<<<<< HEAD
\description{
Creates an empty data frame to be used for mapping spectral data to a
vector of \emph{m/z} values. The vector of \emph{m/z} values is stored in
the first column and can be called by "\code{full_mz}".
=======
Returns a data frame which can be used to map irregularly spaced spectral data to a set range of \emph{m/z} values, contained in the first column of the frame.
=======
\item{min_mz}{Single numeric value; minimum \emph{m/z} value of the
observed range. Default = 53.76.}

\item{max_mz}{Single numeric value; upper end of \emph{m/z} range observed
in spectra. Default = 1100.}

\item{res}{The resolution of peaks; a single numeric value indicating the
step size from the minimum to maximum \emph{m/z} bin. Also known as bin
width. Default = 0.0001.}

\item{dig}{Number of decimal places to round the \emph{m/z} vector to;
required for \emph{m/z} values to match full \emph{m/z} vector. Default =
4.}
>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
=======
\references{
https://github.com/wesleyburr/subMaldi
<<<<<<< HEAD
>>>>>>> db6ced3... Updated documentation
=======
>>>>>>> 270485a... Updated documentation
}
\value{
Returns a data frame which can be used to map irregularly spaced
spectral data to a set range of \emph{m/z} values, contained in the first
column of the frame.
}
<<<<<<< HEAD
<<<<<<< HEAD
=======
\item{min_mz}{Single numeric value; minimum \emph{m/z} value of the
observed range. Default = 53.76.}

\item{max_mz}{Single numeric value; upper end of \emph{m/z} range observed
in spectra. Default = 1100.}

\item{res}{The resolution of peaks; a single numeric value indicating the
step size from the minimum to maximum \emph{m/z} bin. Also known as bin
width. Default = 0.0001.}

\item{dig}{Number of decimal places to round the \emph{m/z} vector to;
required for \emph{m/z} values to match full \emph{m/z} vector. Default =
4.}
}
\value{
Returns a data frame which can be used to map irregularly spaced
spectral data to a set range of \emph{m/z} values, contained in the first
column of the frame.
}
>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
\description{
Creates an empty data frame to be used for mapping spectral data to a
vector of \emph{m/z} values. The vector of \emph{m/z} values is stored in
the first column and can be called by "\code{full_mz}".
<<<<<<< HEAD
=======
\note{
<<<<<<< HEAD
=======
}
>>>>>>> 270485a... Updated documentation
}
<<<<<<< HEAD
>>>>>>> db6ced3... Updated documentation
}
<<<<<<< HEAD
=======
>>>>>>> 834ea85... Updated documentation - removed .tar.gz

<<<<<<< HEAD
<<<<<<< HEAD
=======
>>>>>>> 14c6370... Updated documentation
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
>>>>>>> 270485a... Updated documentation
}
=======

>>>>>>> 5831098... Added pre-processing vignette
=======
>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
=======


>>>>>>> 2931ed6... Added pre-processing vignette
=======
}
>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
\examples{

## Creating an empty spectrum with an m/z range of 500 to 2000 m/z, with a step size of 0.001
spec_df <- createSpecDF(min_mz = 500, max_mz = 2000, res = 0.001, dig = 3)

## Creating an empty spectrum with an m/z range of 100 to 1000, 
# with a step size of 0.0001 and sample names
spec_df <- createSpecDF(min_mz = 100, max_mz = 1000, res = 0.0001, dig = 4)
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD

=======
>>>>>>> 2a95699... Updated documentation
}
\references{
https://github.com/wesleyburr/subMaldi
}
\author{
Kristen Yeh <kristenyeh@trentu.ca>
}
<<<<<<< HEAD
\keyword{array}
\keyword{methods}
=======

\keyword{ methods }
\keyword{ array }
>>>>>>> c6a3769... Updated documentation
=======

}
\references{
https://github.com/wesleyburr/subMaldi
}
\author{
Kristen Yeh <kristenyeh@trentu.ca>
=======
>>>>>>> 834ea85... Updated documentation - removed .tar.gz
}
<<<<<<< HEAD
\keyword{array}
\keyword{methods}
>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
=======

\keyword{ methods }
\keyword{ array }
>>>>>>> 14c6370... Updated documentation
=======

}
\references{
https://github.com/wesleyburr/subMaldi
}
\author{
Kristen Yeh <kristenyeh@trentu.ca>
}
\keyword{array}
\keyword{methods}
>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
