<<<<<<< HEAD
<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_max.R
\name{find_max}
\alias{find_max}
\title{Find Peak Maxima}
\usage{
find_max(dat, mass_dat, spectra_cols)
}
\arguments{
\item{dat}{The name of the spectral data frame, containing \code{m/z} data
in the first column and spectral intensity data in subsequent columns.}

\item{mass_dat}{A character string; the name of the column in \code{dat}
containing the \emph{m/z} data for the spectrum.}

\item{spectra_cols}{A character vector; the names of the column in
\code{dat} containing the intensity data for the spectra-of-interest.}
}
\value{
Returns a data frame indidcating the most intense peaks of each
input spectrum. Indicates the spectrum the data is from, the \emph{m/z}
value associated with the peak, and the intensity of the maxima.
}
\description{
Analyzes spectral data and returns a list of the most intense peak in each
spectrum, including the \emph{m/z} value associated with the peak.
}
\examples{

## Load sample dataset "Master.rda"
data("Master")


## Find maxima of four spectra
find_max(dat = Master, mass_dat = "full_mz", spectra_cols = c("Blank1", "Before1", "After1", "After2")

}
\references{
https://github.com/wesleyburr/subMaldi
}
\author{
Kristen Yeh <kristenyeh@trentu.ca> Wesley Burr <wburr@trentu.ca>
Sophie Castel <sophie.castel@ontariotechu.net>
}
\keyword{methods}
=======
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_max.R
>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
\name{find_max}
\alias{find_max}
\title{Find Peak Maxima}
\usage{
find_max(dat, mass_dat, spectra_cols)
}
\arguments{
\item{dat}{The name of the spectral data frame, containing \code{m/z} data
in the first column and spectral intensity data in subsequent columns.}

\item{mass_dat}{A character string; the name of the column in \code{dat}
containing the \emph{m/z} data for the spectrum.}

\item{spectra_cols}{A character vector; the names of the column in
\code{dat} containing the intensity data for the spectra-of-interest.}
}
\value{
Returns a data frame indidcating the most intense peaks of each
input spectrum. Indicates the spectrum the data is from, the \emph{m/z}
value associated with the peak, and the intensity of the maxima.
}
\description{
Analyzes spectral data and returns a list of the most intense peak in each
spectrum, including the \emph{m/z} value associated with the peak.
}
\examples{

## Load sample dataset "Master.rda"
data("Master")


## Find maxima of four spectra
find_max(dat = Master, mass_dat = "full_mz", spectra_cols = c("Blank1", "Before1", "After1", "After2")
<<<<<<< HEAD
}
<<<<<<< HEAD
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
>>>>>>> 2a95699... Updated documentation
=======

\keyword{ methods }
>>>>>>> c6a3769... Updated documentation
=======

}
\references{
https://github.com/wesleyburr/subMaldi
}
\author{
Kristen Yeh <kristenyeh@trentu.ca> Wesley Burr <wburr@trentu.ca>
Sophie Castel <sophie.castel@ontariotechu.net>
}
\keyword{methods}
>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
