\docType{package}
\name{subMALDI-package}
\alias{subMALDI}
\alias{subMALDI-package}
\title{subMALDI: Organization and Processing of MALDI-MS Datasets}
\description{
To learn more about subMALDI, start with the vignettes:
\code{browseVignettes(package = "subMALDI")}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/wesleyburr/subMALDI/issues}
}

}
\author{
\strong{Maintainer}: Wesley Burr \email{wesleyburr@trentu.ca} (\href{https://orcid.org/0000-0002-2058-1899}{ORCID})

Authors:
\itemize{
  \item Kristen Yeh (\href{https://orcid.org/0000-0002-3411-6816}{ORCID})
  \item Wesley Burr (\href{https://orcid.org/0000-0002-2058-1899}{ORCID})
}
}

